+++ src/main/kotlin/com/opsysinc/learning/data/TagData.kt
@@ -2,7 +2,6 @@ package com.opsysinc.learning.data
 
 import com.opsysinc.learning.util.LRUCache
 import java.util.*
-import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicLong
 
 /**
@@ -15,8 +14,8 @@ data class TagData(val id: String,
                    val tagValue: String,
                    val updatedOn: AtomicLong
                    = AtomicLong(0L),
-                   val hitCtr: AtomicInteger
-                   = AtomicInteger(0),
+                   val hitCtr: AtomicLong
+                   = AtomicLong(0L),
                    val tweets: MutableMap<Long, Long>
                    = Collections.synchronizedMap(LRUCache<Long, Long>(20))
 )
